+++ comp2013/src/main/java/comp2013/View/SnakeFood.java
@@ -93,11 +93,22 @@ public class SnakeFood
 		int randomInt = new Random().nextInt(19);
 		// If a special fruit, set the boolean to true.
 		if(randomInt == 17){
+			// Play the timeline again if it is stopped
+			if(M_Controller.m_View.M_FoodTimeline != null) {
+				M_Controller.m_View.M_FoodTimeline.play();
+			}
 			m_NegativeFruit = true;
 		}
 		else if(randomInt == 18)
 		{
+			if(M_Controller.m_View.M_FoodTimeline != null) {
+				M_Controller.m_View.M_FoodTimeline.play();
+			}
 			m_BonusFruit = true;
+		}
+		// Otherwise stop the timeline for regular fruit
+		else{ if(M_Controller.m_View.M_FoodTimeline != null) {
+			M_Controller.m_View.M_FoodTimeline.stop();}
 		}
 		// Carries out the same as the constructor, but can be called by other functions.
 		this.M_Image = SnakeImageUtil.getImage(String.valueOf(randomInt));
@@ -121,7 +132,7 @@ public class SnakeFood
 	 * @param h2 second image height
 	 * @return Returns true if the image are intersecting, false otherwise.
 	 */
-	private boolean areImagesIntersecting(double x1, double y1, double w1, double h1,
+	public static boolean areImagesIntersecting(double x1, double y1, double w1, double h1,
 										  double x2, double y2, double w2, double h2) {
 		return (x1 < x2 + w2 && x1 + w1 > x2 && y1 < y2 + h2 && y1 + h1 > y2);
 	}